last commit before branching to develop legend. Previou
sly, width and height have been specified for the svg, with the circle diameter shrinking if title is to be displayed above it. Perhaps, the width should be determined similarly, growing if a legend is required (and to accommodate legend text of varying width). So will not worry about width for now, using the circle diameter as a datum and growing sideways as needed. At some point, the final desired width will have to be used to determine the circle diameter before anything is rendered.